connect routes in controller
