+++ controllers/loginController.js
@@ -1,5 +1,5 @@
 const SITE_TITLE = 'Shope';
-module.exports = (request, response) => {
+module.exports.index = (request, response) => {
     // if a user is logged in
     if (request.session.login) {
         // redirect to dashboard
@@ -11,4 +11,17 @@ module.exports = (request, response) => {
             title: 'Login'
         });
     }
+}
+module.exports.submit = (request, response) => {
+    // loop through all users to see if the one of the user from the users array matches with the request body
+    for(var i = 0; i < users.length; i++) {
+        let username = users[i].username;
+        let password = users[i].password;
+        if(request.body.username == username && request.body.password == password) {
+            request.session.login = username;
+            console.log(users[i]);
+            return response.render('login-succes');
+        }
+    }
+    return response.render('login-failed');
 }
