+++ controllers/productsController.js
@@ -1,5 +1,5 @@
 const SITE_TITLE = 'Shope';
-module.exports = (request, response) => {
+module.exports.index = (request, response) => {
     // if a user is logged in
     if (request.session.login) {
         // redirect to dashboard
@@ -9,6 +9,19 @@ module.exports = (request, response) => {
         response.render('product',{
             site_title: SITE_TITLE,
             title: 'Product'
+        });
+    }
+}
+module.exports.details = (request, response) => {
+    // if a user is logged in
+    if (request.session.login) {
+        // redirect to dashboard
+        response.redirect('/dashboard');
+    }else{
+        // render products.html from views
+        response.render('product-details',{
+            site_title: SITE_TITLE,
+            title: 'Product-details'
         });
     }
 }
